Rename moa.tasks.active to moa.tasks.meta
Use the new package moa.tasks.meta for collecting the MetaMainTask
and all tasks that extend it in one place. This way, when looking
for a MetaMainTask via ClassOption.cliStringToObject(), all
subclasses can easily be found.
