+++ moa/src/main/java/moa/tasks/meta/ALPrequentialEvaluationTask.java
@@ -17,7 +17,7 @@
  *    along with this program. If not, see <http://www.gnu.org/licenses/>.
  *    
  */
-package moa.tasks.active;
+package moa.tasks.meta;
 
 import java.awt.Color;
 import java.io.File;
