+++ moa/src/main/java/moa/tasks/meta/ALTaskThread.java
@@ -19,7 +19,7 @@
  *    along with this program. If not, see <http://www.gnu.org/licenses/>.
  *    
  */
-package moa.tasks.active;
+package moa.tasks.meta;
 
 import java.util.List;
